move sbom image logic
